allow records step update only to data analyzer / code cleanup (#1757)
